3o1s/Ai/DigDug/requirements.txt
pillow
requests
pygame
aiohttp
async-timeout
websockets
yarl
